Fixing a couple compiler warnings
git-svn-id: svn+ssh://sxs-archive.tapir.caltech.edu/SpEC/SVNroot/trunk/Support/WaveExtrapolation/Waveforms@18045 aeca8a61-ec13-0410-a8d4-8d6e48c7bcfb

+++ Objects/Waveform.cpp
@@ -973,7 +973,7 @@ public:
     arga.erase(arga.begin(), arga.begin()+i);
   }
   
-  double LM_a() { return LMa; }
+  int LM_a() { return LMa; }
   
   double darg(const double dt, const unsigned int mode) const {
     vector<double> argA = WaveformUtilities::Interpolate(a.T(), a.Arg(mode), t);
